+++ cmd/fofa/cmd/search.go
@@ -98,6 +98,13 @@ var searchCmd = &cli.Command{
 			Usage:       "search result for over a year",
 			Destination: &full,
 		},
+		&cli.IntFlag{
+			Name:        "batchSize",
+			Aliases:     []string{"bs"},
+			Value:       1000,
+			Usage:       "amount of data contained in each page batch",
+			Destination: &batchSize,
+		},
 		&cli.BoolFlag{
 			Name:        "uniqByIP",
 			Value:       false,
@@ -351,6 +358,7 @@ func SearchAction(ctx *cli.Context) error {
 			DeWildcard:  deWildcard,
 			Filter:      filter,
 			DedupHost:   dedupHost,
+			BatchSize:   batchSize,
 		})
 		if err != nil {
 			return err